+++ test/test_fc_layer.cpp
@@ -7,12 +7,12 @@
 
 TEST_CASE( "Test_FullyConnectedLayer" )
 {
-    size_t input_size  = 2;
-    size_t output_size = 3;
-    auto layer         = Robbie::FCLayer<double>( input_size, output_size );
-    auto y_get         = Robbie::Vector<double>( output_size );
-    auto y_actual      = Robbie::Vector<double>( output_size );
-    auto X             = Robbie::Vector<double>( input_size );
+    long input_size  = 2;
+    long output_size = 3;
+    auto layer       = Robbie::FCLayer<double>( input_size, output_size );
+    auto y_get       = Robbie::Vector<double>( output_size );
+    auto y_actual    = Robbie::Vector<double>( output_size );
+    auto X           = Robbie::Vector<double>( input_size );
     X << -3, 2;
 
     auto weights = Robbie::Matrix<double>( input_size, output_size );
